move redesigned classes into separate files
