+++ echopype/convert/__init__.py
@@ -8,6 +8,9 @@ The current version supports:
 - ASL Environmental Sciences AZFP echosounder ``.01A`` data
 """
 from .convert_ui import Convert
-from .ek60 import ConvertEK60
-from .ek80 import ConvertEK80
-from .azfp import ConvertAZFP
+from .ek60 import ParseEK60
+from .ek80 import ParseEK80
+from .azfp import ParseAZFP
+from .set_groups_azfp import SetGroupsAZFP
+from .set_groups_ek60 import SetGroupsEK60
+from .set_groups_ek80 import SetGroupsEK80
