+++ echopype/convert/convert_ui.py
@@ -8,8 +8,12 @@ import xarray as xr
 import netCDF4
 import numpy as np
 import dask
-from .convertbase_new import ParseEK60, ParseEK80, ParseAZFP
-from .utils.setgroups_new import SetGroupsEK60, SetGroupsEK80, SetGroupsAZFP
+from .azfp import ParseAZFP
+from .ek60 import ParseEK60
+from .ek80 import ParseEK80
+from .set_groups_azfp import SetGroupsAZFP
+from .set_groups_ek60 import SetGroupsEK60
+from .set_groups_ek80 import SetGroupsEK80
 
 
 class Convert:
