+++ echopype/convert/set_groups_azfp.py
@@ -1,32 +1,43 @@
-import xarray as xr
-import netCDF4
-import zarr
+"""
+Class to save unpacked echosounder data to appropriate groups in netcdf or zarr.
+"""
 import os
+import xarray as xr
+import numpy as np
+import zarr
+import netCDF4
 from .set_groups_base import SetGroupsBase
 
 
 class SetGroupsAZFP(SetGroupsBase):
-    """Class for setting groups in netCDF file for AZFP data.
+    """Class for saving groups to netcdf or zarr from AZFP data files.
     """
+    def save(self):
+        """Actually save groups to file by calling the set methods.
+        """
+        sonar_values = ('ASL Environmental Sciences', 'Acoustic Zooplankton Fish Profiler',
+                        int(self.convert_obj.unpacked_data['serial_number']),
+                        'Based on AZFP Matlab Toolbox', '1.4', 'echosounder')
+        self.set_toplevel("AZFP")
+        self.set_env()
+        self.set_provenance()
+        self.set_platform()
+        self.set_sonar(sonar_values)
+        self.set_beam()
+        self.set_vendor()
 
-    def set_env(self, env_dict):
-        """Set the Environment group in the AZFP netCDF file.
-        AZFP includes additional variables 'salinity' and 'pressure'
-
-        Parameters
-        ----------
-        env_dict
-            dictionary containing environment group params
-                         env_dict['frequency']
-                         env_dict['absorption_coeff']
-                         env_dict['sound_speed']
+    def set_env(self, ):
+        """Set the Environment group.
         """
         # Only save environment group if file_path exists
-        if not os.path.exists(self.file_path):
+        if not os.path.exists(self.output_path):
             print('netCDF file does not exist, exiting without saving Environment group...')
         else:
-            ds = xr.Dataset({'temperature': (['ping_time'], env_dict['temperature'])},
-                            coords={'ping_time': (['ping_time'], env_dict['ping_time'],
+            # TODO Look at why this cannot be encoded without the modifications
+            ping_time = (self.convert_obj.ping_time - np.datetime64('1970-01-01T00:00:00')) / np.timedelta64(1, 's')
+            # ping_time = self.convert_obj.ping_time
+            ds = xr.Dataset({'temperature': (['ping_time'], self.convert_obj.unpacked_data['temperature'])},
+                            coords={'ping_time': (['ping_time'], ping_time,
                                     {'axis': 'T',
                                      'calendar': 'gregorian',
                                      'long_name': 'Timestamp of each ping',
@@ -34,167 +45,183 @@ class SetGroupsAZFP(SetGroupsBase):
                                      'units': 'seconds since 1970-01-01'})},
                             attrs={'long_name': "Water temperature",
                                    'units': "C"})
-
             # save to file
-            if self.format == '.nc':
-                ds.to_netcdf(path=self.file_path, mode='a', group='Environment')
-            elif self.format == '.zarr':
-                if not self.append_zarr:
-                    ds.to_zarr(store=self.file_path, mode='a', group='Environment')
-                else:
-                    ds.to_zarr(store=self.file_path, mode='a', group='Environment', append_dim='ping_time')
+            if self.save_ext == '.nc':
+                ds.to_netcdf(path=self.output_path, mode='a', group='Environment')
+            elif self.save_ext == '.zarr':
+                ds.to_zarr(store=self.output_path, mode='a', group='Environment')
 
-    def set_platform(self, platform_dict):
-        """Set the Platform group in the AZFP nc file. AZFP does not record pitch, roll, and heave.
-
-        Parameters
-        ----------
-        platform_dict
-            dictionary containing platform parameters
+    def set_platform(self):
+        """Set the Platform group.
         """
-        if not os.path.exists(self.file_path):
+        platform_dict = {'platform_name': self.ui_param['platform_name'],
+                         'platform_type': self.ui_param['platform_type'],
+                         'platform_code_ICES': self.ui_param['platform_code_ICES']}
+        # Only save platform group if file_path exists
+        if not os.path.exists(self.output_path):
             print("netCDF file does not exist, exiting without saving Platform group...")
-        elif self.format == '.nc':
-            with netCDF4.Dataset(self.file_path, 'a', format='NETCDF4') as ncfile:
-                plat = ncfile.createGroup('Platform')
-                [plat.setncattr(k, v) for k, v in platform_dict.items()]
-        elif self.format == '.zarr' and not self.append_zarr:    # Do not save platform if appending
-            zarrfile = zarr.open(self.file_path, mode='a')
-            plat = zarrfile.create_group('Platform')
-            for k, v in platform_dict.items():
-                plat.attrs[k] = v
+        else:
+            if self.save_ext == '.nc':
+                with netCDF4.Dataset(self.output_path, 'a', format='NETCDF4') as ncfile:
+                    plat = ncfile.createGroup('Platform')
+                    [plat.setncattr(k, v) for k, v in platform_dict.items()]
+            elif self.save_ext == '.zarr':
+                zarrfile = zarr.open(self.output_path, mode='a')
+                plat = zarrfile.create_group('Platform')
+                for k, v in platform_dict.items():
+                    plat.attrs[k] = v
 
-    def set_beam(self, beam_dict):
-        """Set the Beam group in the AZFP nc file.
-
-        Parameters
-        ----------
-        beam_dict
-            dictionary containing general beam parameters
+    def set_beam(self):
+        """Set the Beam group.
         """
+        unpacked_data = self.convert_obj.unpacked_data
+        parameters = self.convert_obj.parameters
+        anc = np.array(unpacked_data['ancillary'])   # convert to np array for easy slicing
+        dig_rate = unpacked_data['dig_rate']         # dim: freq
+        freq = np.array(unpacked_data['frequency']) * 1000    # Frequency in Hz
+        ping_time = (self.convert_obj.ping_time - np.datetime64('1900-01-01T00:00:00')) / np.timedelta64(1, 's')
 
-        ds = xr.Dataset({'backscatter_r': (['frequency', 'ping_time', 'range_bin'], beam_dict['backscatter_r']),
-                         'equivalent_beam_angle': (['frequency'], beam_dict['EBA']),
-                         'gain_correction': (['frequency'], beam_dict['gain_correction']),
-                         'sample_interval': (['frequency'], beam_dict['sample_interval'],
+        # Build variables in the output xarray Dataset
+        N = []   # for storing backscatter_r values for each frequency
+        Sv_offset = np.zeros(freq.shape)
+        for ich in range(len(freq)):
+            Sv_offset[ich] = self.convert_obj._calc_Sv_offset(freq[ich], unpacked_data['pulse_length'][ich])
+            N.append(np.array([unpacked_data['counts'][p][ich]
+                               for p in range(len(unpacked_data['year']))]))
+
+        # Largest number of counts along the range dimension among the different channels
+        longest_range_bin = np.max(unpacked_data['num_bins'])
+        range_bin = np.arange(longest_range_bin)
+
+        # Pad power data
+        if any(unpacked_data['num_bins'] != longest_range_bin):
+            N_tmp = np.full((len(N), len(ping_time), longest_range_bin), np.nan)
+            for i, n in enumerate(N):
+                N_tmp[i, :, :n.shape[1]] = n
+            N = N_tmp
+            del N_tmp
+
+        tdn = unpacked_data['pulse_length'] / 1e6  # Convert microseconds to seconds
+        range_samples_xml = np.array(parameters['range_samples'])         # from xml file
+        range_samples_per_bin = unpacked_data['range_samples_per_bin']    # from data header
+
+        # Calculate sample interval in seconds
+        if len(dig_rate) == len(range_samples_per_bin):
+            sample_int = range_samples_per_bin / dig_rate
+        else:
+            raise ValueError("dig_rate and range_samples not unique across frequencies")
+
+        ds = xr.Dataset({'backscatter_r': (['frequency', 'ping_time', 'range_bin'], N),
+                         'equivalent_beam_angle': (['frequency'], parameters['BP']),
+                         'gain_correction': (['frequency'], parameters['gain']),
+                         'sample_interval': (['frequency'], sample_int,
                                              {'units': 's'}),
-                         'transmit_duration_nominal': (['frequency'], beam_dict['transmit_duration_nominal'],
+                         'transmit_duration_nominal': (['frequency'], tdn,
                                                        {'long_name': 'Nominal bandwidth of transmitted pulse',
                                                         'units': 's',
                                                         'valid_min': 0.0}),
-                         'temperature_counts': (['ping_time'], beam_dict['temperature_counts']),
-                         'tilt_x_count': (['ping_time'], beam_dict['tilt_x_count']),
-                         'tilt_y_count': (['ping_time'], beam_dict['tilt_y_count']),
-                         'tilt_x': (['ping_time'], beam_dict['tilt_x']),
-                         'tilt_y': (['ping_time'], beam_dict['tilt_y']),
-                         'cos_tilt_mag': (['ping_time'], beam_dict['cos_tilt_mag']),
-                         'DS': (['frequency'], beam_dict['DS']),
-                         'EL': (['frequency'], beam_dict['EL']),
-                         'TVR': (['frequency'], beam_dict['TVR']),
-                         'VTX': (['frequency'], beam_dict['VTX']),
-                         'Sv_offset': (['frequency'], beam_dict['Sv_offset']),
-                         'number_of_samples_digitized_per_pings': (['frequency'], beam_dict['range_samples']),
+                         'temperature_counts': (['ping_time'], anc[:, 4]),
+                         'tilt_x_count': (['ping_time'], anc[:, 0]),
+                         'tilt_y_count': (['ping_time'], anc[:, 1]),
+                         'tilt_x': (['ping_time'], unpacked_data['tilt_x']),
+                         'tilt_y': (['ping_time'], unpacked_data['tilt_y']),
+                         'cos_tilt_mag': (['ping_time'], unpacked_data['cos_tilt_mag']),
+                         'DS': (['frequency'], parameters['DS']),
+                         'EL': (['frequency'], parameters['EL']),
+                         'TVR': (['frequency'], parameters['TVR']),
+                         'VTX': (['frequency'], parameters['VTX']),
+                         'Sv_offset': (['frequency'], Sv_offset),
+                         'number_of_samples_digitized_per_pings': (['frequency'], range_samples_xml),
                          'number_of_digitized_samples_averaged_per_pings': (['frequency'],
-                                                                            beam_dict['range_averaging_samples'])},
-                        coords={'frequency': (['frequency'], beam_dict['frequency'],
+                                                                            parameters['range_averaging_samples'])},
+                        coords={'frequency': (['frequency'], freq,
                                               {'units': 'Hz',
                                                'valid_min': 0.0}),
-                                'ping_time': (['ping_time'], beam_dict['ping_time'],
+                                'ping_time': (['ping_time'], ping_time,
                                               {'axis': 'T',
                                                'calendar': 'gregorian',
                                                'long_name': 'Timestamp of each ping',
                                                'standard_name': 'time',
-                                               'units': 'seconds since 1970-01-01'}),
-                                'range_bin': (['range_bin'], beam_dict['range_bin'])},
+                                               'units': 'seconds since 1900-01-01'}),
+                                'range_bin': (['range_bin'], range_bin)},
                         attrs={'beam_mode': '',
                                'conversion_equation_t': 'type_4',
-                               'number_of_frequency': beam_dict['number_of_frequency'],
-                               'number_of_pings_per_burst': beam_dict['number_of_pings_per_burst'],
-                               'average_burst_pings_flag': beam_dict['average_burst_pings_flag'],
+                               'number_of_frequency': parameters['num_freq'],
+                               'number_of_pings_per_burst': parameters['pings_per_burst'],
+                               'average_burst_pings_flag': parameters['average_burst_pings'],
                                # Temperature coefficients
-                               'temperature_ka': beam_dict['temperature_ka'],
-                               'temperature_kb': beam_dict['temperature_kb'],
-                               'temperature_kc': beam_dict['temperature_kc'],
-                               'temperature_A': beam_dict['temperature_A'],
-                               'temperature_B': beam_dict['temperature_B'],
-                               'temperature_C': beam_dict['temperature_C'],
+                               'temperature_ka': parameters['ka'],
+                               'temperature_kb': parameters['kb'],
+                               'temperature_kc': parameters['kc'],
+                               'temperature_A': parameters['A'],
+                               'temperature_B': parameters['B'],
+                               'temperature_C': parameters['C'],
                                # Tilt coefficients
-                               'tilt_X_a': beam_dict['tilt_X_a'],
-                               'tilt_X_b': beam_dict['tilt_X_b'],
-                               'tilt_X_c': beam_dict['tilt_X_c'],
-                               'tilt_X_d': beam_dict['tilt_X_d'],
-                               'tilt_Y_a': beam_dict['tilt_Y_a'],
-                               'tilt_Y_b': beam_dict['tilt_Y_b'],
-                               'tilt_Y_c': beam_dict['tilt_Y_c'],
-                               'tilt_Y_d': beam_dict['tilt_Y_d']})
-        n_settings = {}
-        z_settings = {}
-        if self.compress:
-            n_settings = {'backscatter_r': {'zlib': True, 'complevel': 4}}
-            z_settings = {'backscatter_r': {'compressor': zarr.Blosc(cname='zstd', clevel=3, shuffle=2)}}
+                               'tilt_X_a': parameters['X_a'],
+                               'tilt_X_b': parameters['X_b'],
+                               'tilt_X_c': parameters['X_c'],
+                               'tilt_X_d': parameters['X_d'],
+                               'tilt_Y_a': parameters['Y_a'],
+                               'tilt_Y_b': parameters['Y_b'],
+                               'tilt_Y_c': parameters['Y_c'],
+                               'tilt_Y_d': parameters['Y_d']})
 
-        if self.format == '.nc':
-            ds.to_netcdf(path=self.file_path, mode='a', group='Beam', encoding=n_settings)
-        elif self.format == '.zarr':
-            if not self.append_zarr:
-                ds.to_zarr(store=self.file_path, mode='a', group='Beam', encoding=z_settings)
-            else:
-                ds.to_zarr(store=self.file_path, mode='a', group='Beam', append_dim='ping_time')
+        if self.save_ext == '.nc':
+            nc_encoding = {var: self.NETCDF_COMPRESSION_SETTINGS for var in ds.data_vars} if self.compress else {}
+            ds.to_netcdf(path=self.output_path, mode='a', group='Beam', encoding=nc_encoding)
+        elif self.save_ext == '.zarr':
+            zarr_encoding = {var: self.ZARR_COMPRESSION_SETTINGS for var in ds.data_vars} if self.compress else {}
+            ds.to_zarr(store=self.output_path, mode='a', group='Beam', encoding=zarr_encoding)
 
-    def set_vendor_specific(self, vendor_dict):
-        """Set the Vendor-specific group in the AZFP nc file.
-
-        Parameters
-        ----------
-        vendor_dict
-            dictionary containing vendor-specific parameters
+    def set_vendor(self):
+        """Set the Vendor-specific group.
         """
+        unpacked_data = self.convert_obj.unpacked_data
+        freq = np.array(unpacked_data['frequency']) * 1000    # Frequency in Hz
+        ping_time = (self.convert_obj.ping_time - np.datetime64('1900-01-01T00:00:00')) / np.timedelta64(1, 's')
 
         ds = xr.Dataset({
-            'digitization_rate': (['frequency'], vendor_dict['digitization_rate']),
-            'lockout_index': (['frequency'], vendor_dict['lockout_index']),
-            'number_of_bins_per_channel': (['frequency'], vendor_dict['num_bins']),
-            'number_of_samples_per_average_bin': (['frequency'], vendor_dict['range_samples_per_bin']),
-            'board_number': (['frequency'], vendor_dict['board_number']),
-            'data_type': (['frequency'], vendor_dict['data_type']),
-            'ping_status': (['ping_time'], vendor_dict['ping_status']),
-            'number_of_acquired_pings': (['ping_time'], vendor_dict['number_of_acquired_pings']),
-            'first_ping': (['ping_time'], vendor_dict['first_ping']),
-            'last_ping': (['ping_time'], vendor_dict['last_ping']),
-            'data_error': (['ping_time'], vendor_dict['data_error']),
-            'sensor_flag': (['ping_time'], vendor_dict['sensor_flag']),
-            'ancillary': (['ping_time', 'ancillary_len'], vendor_dict['ancillary']),
-            'ad_channels': (['ping_time', 'ad_len'], vendor_dict['ad_channels']),
-            'battery_main': (['ping_time'], vendor_dict['battery_main']),
-            'battery_tx': (['ping_time'], vendor_dict['battery_tx'])},
+            'digitization_rate': (['frequency'], unpacked_data['dig_rate']),
+            'lockout_index': (['frequency'], unpacked_data['lockout_index']),
+            'number_of_bins_per_channel': (['frequency'], unpacked_data['num_bins']),
+            'number_of_samples_per_average_bin': (['frequency'], unpacked_data['range_samples_per_bin']),
+            'board_number': (['frequency'], unpacked_data['board_num']),
+            'data_type': (['frequency'], unpacked_data['data_type']),
+            'ping_status': (['ping_time'], unpacked_data['ping_status']),
+            'number_of_acquired_pings': (['ping_time'], unpacked_data['num_acq_pings']),
+            'first_ping': (['ping_time'], unpacked_data['first_ping']),
+            'last_ping': (['ping_time'], unpacked_data['last_ping']),
+            'data_error': (['ping_time'], unpacked_data['data_error']),
+            'sensor_flag': (['ping_time'], unpacked_data['sensor_flag']),
+            'ancillary': (['ping_time', 'ancillary_len'], unpacked_data['ancillary']),
+            'ad_channels': (['ping_time', 'ad_len'], unpacked_data['ad']),
+            'battery_main': (['ping_time'], unpacked_data['battery_main']),
+            'battery_tx': (['ping_time'], unpacked_data['battery_tx'])},
             coords={
-                'frequency': (['frequency'], vendor_dict['frequency'],
+                'frequency': (['frequency'], freq,
                               {'units': 'Hz',
                                'valid_min': 0.0}),
-                'ping_time': (['ping_time'], vendor_dict['ping_time'],
+                'ping_time': (['ping_time'], ping_time,
                               {'axis': 'T',
                                'calendar': 'gregorian',
                                'long_name': 'Timestamp of each ping',
                                'standard_name': 'time',
                                'units': 'seconds since 1970-01-01'}),
-                'ancillary_len': (['ancillary_len'], vendor_dict['ancillary_len']),
-                'ad_len': (['ad_len'], vendor_dict['ad_len'])},
+                'ancillary_len': (['ancillary_len'], list(range(len(unpacked_data['ancillary'][0])))),
+                'ad_len': (['ad_len'], list(range(len(unpacked_data['ad'][0]))))},
             attrs={
-                'profile_flag': vendor_dict['profile_flag'],
-                'profile_number': vendor_dict['profile_number'],
-                'burst_interval': vendor_dict['burst_interval'],
-                'ping_per_profile': vendor_dict['ping_per_profile'],
-                'average_pings_flag': vendor_dict['average_pings_flag'],
-                'spare_channel': vendor_dict['spare_channel'],
-                'ping_period': vendor_dict['ping_period'],
-                'phase': vendor_dict['phase'],
-                'number_of_channels': vendor_dict['number_of_channels']}
+                'profile_flag': unpacked_data['profile_flag'],
+                'profile_number': unpacked_data['profile_number'],
+                'burst_interval': unpacked_data['burst_int'],
+                'ping_per_profile': unpacked_data['ping_per_profile'],
+                'average_pings_flag': unpacked_data['avg_pings'],
+                'spare_channel': unpacked_data['spare_chan'],
+                'ping_period': unpacked_data['ping_period'],
+                'phase': unpacked_data['phase'],
+                'number_of_channels': unpacked_data['num_chan']}
         )
 
-        if self.format == '.nc':
-            ds.to_netcdf(path=self.file_path, mode='a', group='Vendor')
-        elif self.format == '.zarr':
-            if not self.append_zarr:
-                ds.to_zarr(store=self.file_path, mode='a', group='Vendor')
-            else:
-                ds.to_zarr(store=self.file_path, mode='a', group='Vendor', append_dim='ping_time')
+        if self.save_ext == '.nc':
+            ds.to_netcdf(path=self.output_path, mode='a', group='Vendor')
+        elif self.save_ext == '.zarr':
+            ds.to_zarr(store=self.output_path, mode='a', group='Vendor')
