+++ echopype/tests/test_convert_azfp.py
@@ -3,7 +3,7 @@ import shutil
 import numpy as np
 import xarray as xr
 import pandas as pd
-from ..convert import Convert
+from echopype.convert import Convert
 
 # raw_path = './echopype/data/azfp/17031001.01A'     # Canada (Different ranges)
 # xml_path = './echopype/data/azfp/17030815.XML'     # Canada (Different ranges)
@@ -145,3 +145,5 @@ def test_validate_path():
         raise AssertionError
 
     os.rmdir(directory)
+
+test_convert_raw_echoview()
