+++ echopype/tests/test_process_azfp.py
@@ -25,8 +25,8 @@ def test_process_AZFP_matlab():
     TS_test = xr.open_dataset(azfp_test_TS_path)
 
     # Convert to .nc file
-    tmp_convert = Convert(azfp_01a_path, azfp_xml_path)
-    tmp_convert.raw2nc()
+    tmp_convert = Convert(azfp_01a_path, azfp_xml_path, model='AZFP')
+    tmp_convert.to_netcdf()
 
     tmp_echo = Process(tmp_convert.nc_path)
     tmp_echo.calibrate(save=True)
@@ -58,7 +58,7 @@ def test_process_AZFP_matlab():
 # def test_process_AZFP_echoview():
 #     # Convert to .nc file
 #     tmp_convert = Convert(azfp_01a_path, azfp_xml_path)
-#     tmp_convert.raw2nc()
+#     tmp_convert.to_netcdf()
 
 #     tmp_echo = Process(tmp_convert.nc_path)
 #     tmp_echo.calibrate()
